MINOR First commit
This is the very first commit. Consists only of a simple read-me file in
the markdown language (intended for GitHub). Should be readable
everywhere.

This projects will try to accomplish a simple Firefox version 45 and
later that will add about:reader?url= to the beginning of the URL. This
will make the site be in the "reader" format as some doesn't present it.

This extension has no intention (as of now) to rearrange a page content
to make it readable. It might never have.

For license, contact the creator - me - through e-mail. Find it! :)