MINOR Add icons
These icons are used both in the action button in the address bar or in
any place in the UI. These icons are also used in AMO (the Mozilla's
website for add-ons or whatever).

There are three different sizes: 32x32, 48x48 and 96x96. Measurements in
pixels. PNG format although I plan on using SVG if it isn't complicated.

I'm not satisfied yet with these icons. I'm gonna change it - probably.

They were made with InkScape. A dear tool to me!

Edit: this amends just update the version on manifest file.