MINOR Add background script (add-on's brain)
This script is executed every time the user clicks the add-on's button.
It tries to toggle the Reader Mode (Reader View) on/off. If it does not
succeed, it give the console an error message.

To avoid this error, it is necessary to check if the page is an article
- meaning it can be put under Reader Mode - or if it already is in
  Reader Mode. If the page isn't an article nor is in Reader Mode, then
it gives the user a message saying the page does not support the desired
Reader Mode.

The isArticle isn't exactly about the <article> HTML tag. There isn't
much documentation on it and one needs to see the sources but as far as
my research went along, it should at least have a <p> tag. Or a lot of
text.

This message is done with an alert invoked with an 'injection' script.
This script must be called with executeScript() from Tabs.tabs because
the background script only acts for action buttons. The refeered method
works on the page of the desired tab (the active one, in this case).

The ffreader.js needs permission for "tabs". Thanks to executeScript, it
also needs permissions for "activeTab". Both already on manifest.json.

So, basically:

- check if not an article and not in Reader Mode already;
- if not for both, alert the user;
- otherwise tries to toggle the page on/off from Reader Mode;
- in case of an error, informs through console (for DEVs and people
  alike);

There isn't much more to be done. Maybe - probably not - I will add
suppport for common pages to be set to Reader Mode. I'm not even sure if
it is possible as the original intent was there is a limit to editions
and a truthful lack of documentation on webextentions.

TODO support Chromium derivatives as they do not support the toggle
function that Firefox supports.

PS: This probably needs Firefox 58+ to work.